+++ AvgScript/src/functions/color.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 
 import { lineCommentCache } from '../lib/comment';
-import { currentLineCommand } from '../lib/dialogue';
+import { LineType } from '../lib/dialogue';
 import { commandInfoList, ParamType } from '../lib/dict';
 import { regexHexColor, regexRep } from '../lib/regExp';
 import { getAllParams } from '../lib/utilities';
@@ -71,7 +71,7 @@ export const colorProvider = vscode.languages.registerColorProvider('AvgScript',
 					}
 				};
 
-				if (currentLineCommand(text)) {
+				if (lineInfo.lineType === LineType.command) {
 					const command = params[0].substring(1);
 					const paramDefinition = commandInfoList.getValue(command);