+++ AvgScript/src/functions/label.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 
 import { CacheInterface } from '../lib/cacheInterface';
 import { currentLineNotComment, lineCommentCache } from '../lib/comment';
-import { currentLineCommand, currentLineLabel } from '../lib/dialogue';
+import { LineType } from '../lib/dialogue';
 import { InlayHintType, commandInfoList } from '../lib/dict';
 import { regexRep } from '../lib/regExp';
 import { getAllParams } from "../lib/utilities";
@@ -128,12 +128,12 @@ export const labelDefinition = vscode.languages.registerDefinitionProvider('AvgS
 				return undefined;
 			}
 
-			let { line, lineStart, linePrefix, curPos } = parseCommentResult;
+			let { line, lineStart, linePrefix, lineType, curPos } = parseCommentResult;
 
 			const params = getAllParams(line);
 			const paramNum = params.length - 1;
 
-			if (currentLineCommand(line)) {
+			if (lineType === LineType.command) {
 				const command = params[0].substring(1);
 				const paramDefinition = commandInfoList.getValue(command);
 
@@ -192,13 +192,13 @@ export const labelReference = vscode.languages.registerReferenceProvider('AvgScr
 				return undefined;
 			}
 
-			let { line, lineStart, linePrefix, curPos } = parseCommentResult;
+			let { line, lineStart, linePrefix, lineType, curPos } = parseCommentResult;
 
-			if (currentLineLabel(line)) {
+			if (lineType !== LineType.label) {
 				return undefined;
 			}
 
-			let label = line.substring(line.indexOf(";") + 1);
+			let label = line.substring(1);
 
 			lineCommentCache.iterateDocumentCacheWithoutComment(document, (lineInfo) => {
 				let text = lineInfo.textNoCommentAndLangPrefix;
@@ -209,7 +209,7 @@ export const labelReference = vscode.languages.registerReferenceProvider('AvgScr
 				const params = getAllParams(text);
 				const paramNum = params.length - 1;
 
-				if (currentLineCommand(text)) {
+				if (lineInfo.lineType === LineType.command) {
 					const command = params[0].substring(1);
 					const paramDefinition = commandInfoList.getValue(command);