+++ AvgScript/src/functions/preview.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 
 import { activeEditor } from '../extension';
 import { lineCommentCache } from '../lib/comment';
-import { currentLineDialogue } from '../lib/dialogue';
+import { LineType } from '../lib/dialogue';
 import { confPreview_AlwaysSendingMessage } from './command';
 import { basePath, basePathUpdated, projectConfig } from './file';
 
@@ -115,9 +115,9 @@ export class Previewer {
 			if (curCache.comment[lineNumber]) { continue; }
 
 			const parseResult = curCache.result[lineNumber]!;
-			let { line, lineStart, linePrefix, curPos } = parseResult;
+			let { line, lineStart, linePrefix, lineType, curPos } = parseResult;
 
-			const bCurLineText = currentLineDialogue(line);
+			const bCurLineText = lineType === LineType.dialogue;
 
 			if (lineNumber >= cursorAt) {
 				bReached = true;